+++ modules/graphql_1/main.tf
@@ -61,7 +61,6 @@ data "aws_iam_policy_document" "gql_task_execution_policy_document" {
     ]
     resources = [
       var.db_connect_info_secret_arn,
-      var.s3_access_credentials_secret_arn
     ]
   }
 }
@@ -240,10 +239,6 @@ resource "aws_ecs_task_definition" "gql_task" {
       {
         "name": "DB_CONNECT_INFO",
         "valueFrom": "${var.db_connect_info_secret_arn}"
-      },
-      {
-        "name": "S3_ACCESS_CREDENTIALS",
-        "valueFrom": "${var.s3_access_credentials_secret_arn}"
       }
     ],
     "logConfiguration": {